requirements.txt
python-dotenv~=1.0.1
uvicorn[standard]~=0.32.0
websockets~=14.1
starlette~=0.41.2
sse-starlette~=2.1.3
fastapi~=0.115.5
pydantic~=2.9.2
pydantic-settings~=2.5.2
email-validator~=2.2.0
pyjwt~=2.9.0
passlib~=1.7.4
python-multipart~=0.0.17
neo4j~=5.26.0
neomodel~=5.4.0
aiofiles~=24.1.0
mypy~=1.13.0